Only download files that have changed
Some servers provide an etag which is a unique identifier for the
content being served (not always a hash).  This adds two new column to
the archival, an etag and a last-modified field.  If the etag is
available in the archival, and we get the same etag from the request,
then we will not download the file.  The last-modified field is saved
for sanity, it isn't yet used with If-Modified-Since in the request.

This requires a migration:

    alter table archival add column etag character varying;
    alter table archival add column last-modified character varying;

Notification of package updates is only done if at least 1 resource was
downloaded, to avoid unnecessarily running QA and Packagezip.